Реализована возможность удалять задачи 	новый файл:    webapp/templates/delete.html 	изменено:      webapp/templates/task_view.html 	изменено:      webapp/urls.py 	изменено:      webapp/views.py

+++ webapp/views.py
@@ -38,11 +38,12 @@ def task_view(request, pk):
         return HttpResponseNotFound('<h1>Страница не найдена</h1>')
     return render(request, 'task_view.html', {'task': task})
 
-def delete_task(request):
-    pk = request.GET.get('pk')
-    task = Task.objects.get(pk=pk)
-    removal = task.delete()
-    return render(request, 'index.html', {'action': removal})
+def delete_task(request, pk):
+    task = get_object_or_404(Task, pk=pk)
+    if request.method == 'GET':
+        return render(request, 'delete.html', {'task': task})
+    task.delete()
+    return redirect('index')
 
 def edit_task(request, pk):
     task = get_object_or_404(Task, pk=pk)
